requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
h3==4.2.2
Markdown==3.8
pillow==11.2.1
PyJWT==2.9.0
sqlparse==0.5.3
uuid6==2024.7.10
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
environ==1.0
h3==4.2.2
Markdown==3.8
pillow==11.2.1
PyJWT==2.9.0
sqlparse==0.5.3
uuid6==2024.7.10
